% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoxy.R
\docType{data}
\name{epoxy}
\alias{epoxy}
\title{EPOXY data set}
\format{
A data frame with 12 rows and 3 columns:
\describe{
  \item{EXP-TIME}{Exposure time in corrosive substance}
  \item{SYSTEM}{One of four epoxy coating systems}
  \item{CORRATE}{Corrosion rate (nanoamperes per square cm)}
  ...
}
}
\source{
<https://www.amazon.com/Statistics-Engineering-Sciences-William...
}
\usage{
epoxy
}
\description{
A data set of steel panels coated in one of four anti-corrosive epoxy systems and how long they were exposed to corrosive material.
}
\keyword{datasets}
