% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gginterval.R
\name{gginterval}
\alias{gginterval}
\title{gginterval}
\usage{
gginterval(df, alpha = 0.05)
}
\arguments{
\item{df}{data frame}

\item{alpha}{alpha level}
}
\value{
a ggplot and named list of confidence intervals
}
\description{
A function which produces a ggplot and a list of confidence intervals for the rebased data
}
\examples{
{gginterval(epoxy, alpha = 0.05)}
}
