% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRebase.R
\name{myRebase}
\alias{myRebase}
\title{myRebase}
\usage{
myRebase(df, SYSTEM_base, EXP_base)
}
\arguments{
\item{df}{a data frame}

\item{SYSTEM_base}{new base for System}

\item{EXP_base}{new base for Exptime}
}
\value{
a data frame with rebased levels of factors
}
\description{
A function which will rebase a data frame's factors
}
\examples{
\dontrun{myRebase(epoxy, 2, 60)}
}
