% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyepoxy.R
\name{shinyepoxy}
\alias{shinyepoxy}
\title{shinyepoxy}
\usage{
shinyepoxy()
}
\value{
a plot of histograms and a list of confidence intervals
}
\description{
An interactive app with sliders to rebase the levels of the independent variables
}
\examples{
\dontrun{shinyboots()}
}
