% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrebase.R
\name{myrebase}
\alias{myrebase}
\title{myrebase}
\usage{
myrebase(df, SYSTEM_base, EXP_base)
}
\arguments{
\item{df}{data frame `Epoxy`}

\item{SYSTEM_base}{rebase level for `SYSTEM` variable}

\item{EXP_base}{rebase level for `EXPTIME` variable}
}
\value{
a rebased data frame
}
\description{
A fuction which will rebase the levels of the varibles of the `EPOXY` data set
}
\examples{
\dontrun{myrebase(epoxy, 2, 60)}
}
